==> 01_gather_data.ipynb <==
import praw
from psaw import PushshiftAPI
import pandas as pd
import pprint

reddit = praw.Reddit(client_id='4hzFNLs3DXmWcA',
                     client_secret='0JO0s7Nn8bolJYme3wjJmB_Zwm8',
                     user_agent='script:4hzFNLs3DXmWcA:1.0.0 (by /u/SomeGuyNamedGuy)')
api = PushshiftAPI(reddit)

subreddit_ph = reddit.subreddit('Philippines+SafeSexPH')
subreddit_ph

# Make a query string containing the keywords related to contraception that are included in the file
keywords_file = open('helper_files/contraception_keywords.txt', 'r')
queries_list = [('title:"' + word.replace('\n', '') + '" OR selftext:"' + word.replace('\n', '') + '"') for word in keywords_file]
queries1 = ' OR '.join(queries_list[:7])
queries2 = ' OR '.join(queries_list[7:])
queries1

# Search posts in the two subreddits that contain the specified keywords
post_list1 = subreddit_ph.search(query=queries1, limit=100)
post_list1

post_list2 = subreddit_ph.search(query=queries2, limit=100)
post_list2

#for post in post_list2:
#   print('adsfasd')

## Save as a DataFrame

# Attributes per comment: author, body, depth, parent_id, id, score, subreddit_name_prefixed'
df = pd.DataFrame(columns='id submission_title submission_text author body depth parent_id score subreddit_name_prefixed'.split())

for post in post_list1:
    post.comments.replace_more(limit=0)
    comments = post.comments.list()
    for comment in comments:
        data = [comment.id, post.title, post.selftext, comment.author,
               comment.body, comment.depth, comment.parent_id,
               comment.score, comment.subreddit_name_prefixed]
        df.loc[len(df)] = data

for post in post_list2:
    post.comments.replace_more(limit=0)
    comments = post.comments.list()
    for comment in comments:
        data = [comment.id, post.title, post.selftext, comment.author,
               comment.body, comment.depth, comment.parent_id,
               comment.score, comment.subreddit_name_prefixed]
        df.loc[len(df)] = data

df.tail()

df.to_csv('datasets/dataset.csv', index=False)
