==> 02_preprocessing_text.ipynb <==
import re
import pandas as pd
import numpy as np
from nltk.corpus import stopwords
from textblob import TextBlob as tb
from contractions import CONTRACTION_MAP
stopwords = stopwords.words('english')
stopwords.remove('no')
stopwords.remove('not')
stopwords.append('sa')
stopwords.append('na')
stopwords.append('ng')
stopwords.append('ko')
stopwords.append('ba')

data = pd.read_csv('datasets/final_dataset.csv',index_col=None,header=0)
n = data.shape[0]

data.head()

## Cleaning

def clean_text(df):
    """
    Function for cleaning text based on class jupyter notebook on text manipulation
    """
    text = df.copy()
    text['cleaned'] = text['body'].fillna('')
    text['cleaned'] = text['cleaned'].str.lower()
    text['cleaned'] = text['cleaned'].str.replace(r'([-])', ' ')
    text['cleaned'] = text['cleaned'].str.replace(r'([^A-Z a-z\'])|', '')
    text['cleaned'] = text['cleaned'].apply(lambda x: ' '.join([w for w in x.split() if w not in (stopwords)]))
    
    text['cleaned_sub'] = text['submission_text'].fillna('')
    text['cleaned_sub'] = text['cleaned_sub'].str.lower()
    text['cleaned_sub'] = text['cleaned_sub'].str.replace(r'([-])', ' ')
    text['cleaned_sub'] = text['cleaned_sub'].str.replace(r'([^A-Z a-z\'])|', '')
    text['cleaned_sub'] = text['cleaned_sub'].apply(lambda x: ' '.join([w for w in x.split() if w not in (stopwords)]))
    
    return text

clean = clean_text(data)

def expand_contractions(text, contraction_mapping=CONTRACTION_MAP):
    
    contractions_pattern = re.compile('({})'.format('|'.join(contraction_mapping.keys())), 
                                      flags=re.IGNORECASE|re.DOTALL)
    def expand_match(contraction):
        match = contraction.group(0)
        first_char = match[0]
        expanded_contraction = contraction_mapping.get(match)\
                                if contraction_mapping.get(match)\
                                else contraction_mapping.get(match.lower())                       
        expanded_contraction = first_char+expanded_contraction[1:]
        return expanded_contraction
        
    expanded_text = contractions_pattern.sub(expand_match, text)
    expanded_text = re.sub("'", "", expanded_text)
    return expanded_text

expand_list = []
for i in range(n):
    expand_list.append(expand_contractions(clean.loc[i]['cleaned']))
clean['expand'] = expand_list
clean.head()

expand_list = []
for i in range(n):
    expand_list.append(expand_contractions(clean.loc[i]['cleaned_sub']))
clean['expand_sub'] = expand_list
clean.head()

# cleaning again

def clean_text_2(clean):
    text = clean.copy()
    text['expand'] = text['expand'].str.replace(r'([\'])', '')
    text['expand'] = text['expand'].apply(lambda x: ' '.join([w for w in x.split() if w not in (stopwords)]))
    
    text['expand_sub'] = text['expand_sub'].str.replace(r'([\'])', '')
    text['expand_sub'] = text['expand_sub'].apply(lambda x: ' '.join([w for w in x.split() if w not in (stopwords)]))
    return text

expand = clean_text_2(clean)
expand.head()

## Lemmatizing

expand.head(5)

# cleaned
def lemmatize(clean):
    to_lemm = clean.copy()
    to_lemm['lemmatized'] = ''
    for i in range(n):
        lemmatized = ''
        if( len(to_lemm['expand'][i]) > 0 ):
            blob = tb(to_lemm['expand'][i]).sentences[0]
            tags = blob.tags
            for j in range(0,len(tags)):
                if tags[j][1].startswith('NN'):
                    lemmatized += tags[j][0].lemmatize('n')
                    lemmatized += ' '
                elif tags[j][1].startswith('VB'):
                    lemmatized += tags[j][0].lemmatize('v')
                    lemmatized += ' '
                elif tags[j][1].startswith('JJ'):
                    lemmatized += tags[j][0].lemmatize('a')
                    lemmatized += ' '
                else:
                    lemmatized += tags[j][0]
                    lemmatized += ' '
        to_lemm['lemmatized'][i] = lemmatized.strip()
    return to_lemm

lemm = lemmatize(expand)
lemm.head()

# cleaned_sub
def lemmatize_2(clean):
    to_lemm = clean.copy()
    to_lemm['lemm_sub'] = ''
    for i in range(n):
        lemmatized = ''
        if( len(to_lemm['expand_sub'][i]) > 0 ):
            blob = tb(to_lemm['expand_sub'][i]).sentences[0]
            tags = blob.tags
            for j in range(0,len(tags)):
                if tags[j][1].startswith('NN'):
                    lemmatized += tags[j][0].lemmatize('n')
                    lemmatized += ' '
                elif tags[j][1].startswith('VB'):
                    lemmatized += tags[j][0].lemmatize('v')
                    lemmatized += ' '
                elif tags[j][1].startswith('JJ'):
                    lemmatized += tags[j][0].lemmatize('a')
                    lemmatized += ' '
                else:
                    lemmatized += tags[j][0]
                    lemmatized += ' '
        to_lemm['lemm_sub'][i] = lemmatized.strip()
    return to_lemm

lemm = lemmatize_2(lemm)
lemm.head()

def analyze_text(lemm):
    """
    Function for analyzing sentiment and subjectivity based on class jupyter notebook on text analysis
    """
    text = lemm.copy()
    sentiment = []
    subjectivity = []
    for i in text['lemmatized']:
        blob = tb(i)
        sentiment.append(blob.polarity)
        subjectivity.append(blob.subjectivity)
    text['sentiment'] = sentiment
    text['subjectivity'] = subjectivity
    sentiment = []
    subjectivity = []
    for i in text['lemm_sub']:
        blob = tb(i)
        sentiment.append(blob.polarity)
        subjectivity.append(blob.subjectivity)
    text['sentiment_sub'] = sentiment
    text['subjectivity_sub'] = subjectivity
    return text

sent = analyze_text(lemm)
sent.head()

new_df = sent[['id','depth','parent_id','score','lemmatized','lemm_sub','sentiment','subjectivity','sentiment_sub','subjectivity_sub','labels']]

new_df.head()

new_df.to_csv('datasets/preprocessed_data_2.csv')
