==> 03_parent_and_submission.ipynb <==
## Creating a CSV with labels, body text, parent text, and score

import numpy as np
import pandas as pd

data = pd.read_csv('datasets/preprocessed_data_2.csv',index_col=None,header=0)
data = data[['id','depth','parent_id','score','lemmatized','lemm_sub','labels']]

data.head()
data['labels'][377] = 'denying'
data = data.dropna()
data = data.reset_index()

null = data[data.labels.isnull()]
null

data['lemmatized'][data['id']=='d4xury6'][0]

dir_parent = []
for i in range(data.shape[0]):
    if(data['depth'][i] == 0):
        dir_parent.append(data['lemm_sub'][i])
    else:
        parent_id = data['parent_id'][i][3:]
        if( len(data['lemmatized'][data['id']==parent_id]) > 0 ):
            dir_parent.append(list(data['lemmatized'][data['id']==parent_id])[0])
        else:
            dir_parent.append(' ')
data['direct'] = dir_parent
data.head()

df = data[['labels','lemmatized','direct','score']]
df.columns = ['labels','body','parent','score']

df.to_csv('datasets/labelled_dataset.csv')

df.head()
