==> 04_vader_sentiment.ipynb <==
!pip3 install vaderSentiment

import sys
sys.path.append('/usr/local/lib/python3.6/site-packages')

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
import pandas as pd

analyzer = SentimentIntensityAnalyzer()
full_data = pd.read_csv('datasets/final_dataset.csv')
data = pd.read_csv('datasets/preprocessed_data_2.csv')
full_data.head(2)

# Drop the subjectivity columns since that is only for TextBlob

data.drop('subjectivity', axis=1, inplace=True)
data.drop('subjectivity_sub', axis=1, inplace=True)
data.head(3)

data['sentiment_sub'] = full_data.apply(
    lambda row : analyzer.polarity_scores(row['submission_text'])['compound'], axis=1
)
data['sentiment_sub'].head()

data['sentiment'] = full_data.apply(lambda row : analyzer.polarity_scores(row['body'])['compound'], axis=1)
data['sentiment'].head()

data.head(3)

data.to_csv('datasets/preprocessed_data_vader.csv')
